Add product details fragment and navigation component with animation

+++ app/src/main/java/android/ptc/com/ptcflixing/data/model/Product.kt
@@ -15,5 +15,5 @@ data class Product(
     val specialPrice: Int,
     val image: String,
     @SerialName("rating_average")
-    val ratingAverage: Int = 0
+    val ratingAverage: Float = 0F
 )
